### Description (Describe **what** was changed and don't skimp on the **how**)
Fix new clippy error since the upgrade to +nightly-2022-11-03, in the [last docker RP](https://msazure.visualstudio.com/One/_git/Networking-DDoS-Datapath-Docker/pullrequest/7199081?_a=files)

### Checklist
  - [ ] Code changes are covered by Unit Tests
  - [ ] Code changes are covered by Integration Tests

Related work items: #16499088

+++ crates/rte/src/flags.rs
@@ -177,8 +177,8 @@ bitflags! {
         const L2TPV3                = ffi::ETH_RSS_L2TPV3 as u64;
         const PFCP                  = ffi::ETH_RSS_PFCP as u64;
         const PPPOE                 = ffi::ETH_RSS_PPPOE as u64;
-        const ECPRI                 = ffi::ETH_RSS_ECPRI as u64;
-        const MPLS                  = ffi::ETH_RSS_MPLS as u64;
+        const ECPRI                 = ffi::ETH_RSS_ECPRI;
+        const MPLS                  = ffi::ETH_RSS_MPLS;
     }
 }
